backend/pyproject.toml
[project]
name = "secondorder"
version = "0.1.0"
description = "Deep Research Tool powered by Anthropic Claude"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "anthropic>=0.42.0",
    "tavily-python>=0.5.0",
    "httpx>=0.28.0",
    "supabase>=2.11.0",
    "pydantic-settings>=2.7.0",
    "sse-starlette>=2.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.34.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
